tango integration: workaround for pytango bug.
PyTango cannot handle DevULong64 properly.

Reviewed-on: https://trac.frm2.tum.de/review/3101
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Tested-by: Alexander Lenz <alexander.lenz@frm2.tum.de>

+++ lib/nicos/devices/tango.py
@@ -383,7 +383,7 @@ class DigitalOutput(PyTangoDevice, Moveable):
         return self._dev.value
 
     def doStart(self, value):
-        self._dev.value = value
+        self._dev.value = long(value)
 
 
 class StringIO(PyTangoDevice, Device):
